H3: move ErrorCode to proto
