+++ quinn-h3/src/lib.rs
@@ -23,9 +23,8 @@ mod frame;
 mod streams;
 
 use err_derive::Error;
-use quinn::VarInt;
 
-use proto::frame::SettingsFrame;
+use proto::{frame::SettingsFrame, ErrorCode};
 
 pub type Settings = SettingsFrame;
 
@@ -104,43 +103,6 @@ fn try_take<T>(item: &mut Option<T>, msg: &'static str) -> Result<T, Error> {
 
 /// TLS ALPN value for H3
 pub const ALPN: &[u8] = b"h3-20";
-
-#[derive(Copy, Clone, Eq, PartialEq, Debug)]
-pub struct ErrorCode(u32);
-
-macro_rules! error_codes {
-    {$($name:ident = $val:expr,)*} => {
-        impl ErrorCode {
-            $(pub const $name: ErrorCode = ErrorCode($val);)*
-        }
-    }
-}
-
-error_codes! {
-    NO_ERROR = 0x100,
-    GENERAL_PROTOCOL_ERROR = 0x101,
-    INTERNAL_ERROR = 0x102,
-    STREAM_CREATION_ERROR = 0x103,
-    CLOSED_CRITICAL_STREAM = 0x104,
-    FRAME_UNEXPECTED = 0x105,
-    FRAME_ERROR = 0x106,
-    EXCESSIVE_LOAD = 0x107,
-    ID_ERROR = 0x108,
-    SETTINGS_ERROR = 0x109,
-    MISSING_SETTINGS = 0x10A,
-    REQUEST_REJECTED = 0x10B,
-    REQUEST_CANCELLED = 0x10C,
-    REQUEST_INCOMPLETE = 0x10D,
-    EARLY_RESPONSE = 0x10E,
-    CONNECT_ERROR = 0x10F,
-    VERSION_FALLBACK = 0x110,
-}
-
-impl From<ErrorCode> for VarInt {
-    fn from(error: ErrorCode) -> VarInt {
-        error.0.into()
-    }
-}
 
 impl From<frame::Error> for (ErrorCode, String) {
     fn from(err: frame::Error) -> Self {
